feat: implement comprehensive role-based access control (RBAC)

+++ server/src/controllers/blog.controller.js
@@ -1,10 +1,11 @@
 import Blog from '../models/Blog.js';
 import * as BlogService from '../services/blog.service.js';
 import { AppError } from '../utils/AppError.js';
+import { hasPermission } from '../config/roles.js';
 import { asyncHandler } from '../utils/asyncHandler.js';
 
 export const list = asyncHandler(async (req, res) => {
-  res.json({ success: true, data: await BlogService.listBlogs(req.query) });
+  res.json({ success: true, data: await BlogService.listBlogs(req.query, req.user) });
 });
 
 export const get = asyncHandler(async (req, res) => {
@@ -24,6 +25,11 @@ export const update = asyncHandler(async (req, res) => {
 });
 
 export const remove = asyncHandler(async (req, res) => {
+  const blog = await Blog.findById(req.params.id);
+  if (!blog) throw new AppError('Blog not found', 404);
+  if (!hasPermission(req.user.role, 'blog:*') && blog.author.toString() !== req.user.id.toString()) {
+    throw new AppError('You can only delete your own blog', 403);
+  }
   await Blog.findByIdAndDelete(req.params.id);
   res.status(204).send();
 });
